+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.0.25 (Unreleased)
 
+- [#231](https://github.com/influxdata/clockface/pull/231): Introduce `RangeSlider` input component
 - [#229](https://github.com/influxdata/clockface/pull/229): Introduce `SelectableCard` as a port of `CardSelect` from the InfluxDB repository
 - [#224](https://github.com/influxdata/clockface/pull/224): Introduce `Tabs` family of navigation components
 - [#221](https://github.com/influxdata/clockface/pull/221): [Breaking] Rename `ComponentSpacer` to `FlexBox` and `ComponentSpacerFlexChild` to `FlexBoxChild`
@@ -16,7 +17,7 @@
 
 ### 0.0.23
 
-- ¯\_(ツ)_/¯
+- ¯\_(ツ)\_/¯
 
 ### 0.0.22